+++ bison/src/location.c
@@ -173,7 +173,7 @@ location_caret (FILE *out, location loc)
   /* If the line we want to quote is seekable (the same line as the previous
      location), just seek it. If it was before, we lost track of it, so
      return to the start of file.  */
-  if (caret_info.line <= loc.start.line)
+  if (caret_info.line <= (size_t)loc.start.line)
     fseek (caret_info.source, caret_info.offset, SEEK_SET);
   else
     {
@@ -183,7 +183,7 @@ location_caret (FILE *out, location loc)
     }
 
   /* Advance to the line's position, keeping track of the offset.  */
-  while (caret_info.line < loc.start.line)
+  while (caret_info.line < (size_t)loc.start.line)
     caret_info.line += fgetc (caret_info.source) == '\n';
   caret_info.offset = ftell (caret_info.source);
